adding middleware + fixing codeController

+++ backend-code-editor/app/Http/Controllers/CodeController.php
@@ -23,25 +23,33 @@ class CodeController extends Controller
 
     public function deleteCode($id){
         $code = Code::find($id);
-        $code->delete();
-        return response()->json([
-            'code' => 'deleted successfully'
-        ], 204);
-    }
+        if($code){
+            $code->delete();
+            return response()->json([
+                'code' => 'deleted successfully'
+            ], 200);
+        } else {
+            return response()->json(['message' => 'Code snippet not found'], 404);
+        }
+        }
+
+    
 
     public function createCode(Request $req){
         $validated_data = $req->validate([
             'user_id' => 'required|numeric|exists:users,id',
             'file_name' => 'required|string',
+            'code' => 'required|string'
         ]);
-        $code = code::created($validated_data);
+        $code = Code::create($validated_data);
         return response()->json([
+            'status' => 'success',
             'code' => $code
         ], 201);
     }
 
     public function updateCode(Request $req, $id){
-        $code = code::find($id);
+        $code = Code::find($id);
         if($code){
             $validated_data = $req->validate([
                 'user_id' => 'required|numeric|exists:users,id',
@@ -50,6 +58,6 @@ class CodeController extends Controller
             ]);
             $code->update($validated_data);
         }
-        return response()->json(['code' => 'updated successfully']);
+        return response()->json(['code' => 'updated successfully'],204);
     }
 }
